Refactor project path schema
