+++ config_sample.ini
@@ -2,7 +2,5 @@
 token = WRITE_YOUR_DISCORDBOT_TOKEN_HERE
 
 [channels]
-id_announcements = ANNOUNCEMENTS_CHANNEL_ID
 id_gamenewsit = GAMENEWSIT_CHANNEL_ID
 id_gamenewsen = GAMENEWSEN_CHANNEL_ID
-id_testbot = TESTBOT_CHANNEL_ID
