+++ discordbot/events/member_events.py
@@ -1,0 +1,15 @@
+from discord.ext import commands
+
+class MemberEvents(commands.Cog):
+    def __init__(self, bot):
+        self.bot = bot
+
+    @commands.Cog.listener()
+    async def on_member_join(self, member):
+        guild = member.guild
+        if guild.system_channel is not None:
+            to_send = f'Welcome {member.mention} to {guild.name}!'
+            await guild.system_channel.send(to_send)
+
+async def setup(bot):
+    await bot.add_cog(MemberEvents(bot))
